requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.15
django-allauth==0.51.0
django-environ==0.7.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.12
